Revert "Dial back bed preheating temp for PEI on glass"
This reverts commit 2d58333a91c663b8c5637ee53bc796d09a087873.

After switching from PEI on glass (good adhesion, but nightmarish print
removal) to PEI on spring steel, I had to shim the center of my warped
Ender 3 bed with aluminum foil tape to tram the bed successfully.
Unfortunately, this means the center of the bed is a bit cooler than it
used to be. (The spring steel heats up much quicker, but doesn't seem to
have as much thermal mass as the glass.) Increasing the bed temp by 5c
to accommodate.